refactor: Split up AreaCoordinates file.
