+++ src/hooks/useFetchModelAreas.ts
@@ -1,0 +1,11 @@
+import { useQuery } from '@tanstack/react-query';
+import { ModelAreaTypeService } from '../api/generated';
+
+export const useFetchModelAreas = () => {
+  const query = useQuery({
+    queryKey: ['model-area'],
+    queryFn: () => ModelAreaTypeService.getApiModelareatype(),
+  });
+
+  return query;
+};
